+++ django_chatter/tests/test_consumers.py
@@ -1,5 +1,6 @@
 from channels.testing import WebsocketCommunicator
 from channels.routing import ProtocolTypeRouter, URLRouter
+from channels.layers import get_channel_layer
 
 from django.conf import settings
 from django.contrib.auth import get_user_model
@@ -23,10 +24,7 @@ TEST_CHANNEL_LAYERS = {
     },
 }
 
-@pytest.mark.asyncio
-@pytest.mark.django_db(transaction=True)
-async def test_single_tenant_chat_consumer():
-    settings.CHANNEL_LAYERS = TEST_CHANNEL_LAYERS
+def prepare_room_and_user():
     set_up_data()
     room = Room.objects.create()
     user = get_user_model().objects.get(username="user0")
@@ -34,8 +32,15 @@ async def test_single_tenant_chat_consumer():
     room.save()
     client = Client()
     client.force_login(user=user)
+    return client, room, user
+
+@pytest.mark.asyncio
+@pytest.mark.django_db(transaction=True)
+async def test_single_tenant_chat_consumer():
+    settings.CHANNEL_LAYERS = TEST_CHANNEL_LAYERS
+    client, room, user = prepare_room_and_user()
     communicator = WebsocketCommunicator(
-        application, f"/ws/chat/{room.id}/",
+        application, f"/ws/django_chatter/chatrooms/{room.id}/",
         headers=[
             (
                 b'cookie',
@@ -46,7 +51,7 @@ async def test_single_tenant_chat_consumer():
     connected, subprotocol = await communicator.connect()
     assert connected
     data = {
-        'type': 'text',
+        'message_type': 'text',
         'message': "Hello!",
         'sender': user.username,
         'room_id': str(room.id),
@@ -62,15 +67,9 @@ async def test_single_tenant_chat_consumer():
 @pytest.mark.django_db(transaction=True)
 async def test_multitenant_chat_consumer():
     settings.CHANNEL_LAYERS = TEST_CHANNEL_LAYERS
-    set_up_data()
-    room = Room.objects.create()
-    user = get_user_model().objects.get(username="user0")
-    room.members.add(user)
-    room.save()
-    client = Client()
-    client.force_login(user=user)
+    client, room, user = prepare_room_and_user()
     communicator = WebsocketCommunicator(
-        multitenant_application, f"/ws/chat/{room.id}/",
+        multitenant_application, f"/ws/django_chatter/chatrooms/{room.id}/",
         headers=[
             (
                 b'cookie',
@@ -81,7 +80,7 @@ async def test_multitenant_chat_consumer():
     connected, subprotocol = await communicator.connect()
     assert connected
     data = {
-        'type': 'text',
+        'message_type': 'text',
         'message': "Hello!",
         'sender': user.username,
         'room_id': str(room.id),
@@ -107,15 +106,9 @@ async def test_multitenant_chat_consumer():
 @pytest.mark.django_db(transaction=True)
 async def test_harmful_message_in_chat_consumer():
     settings.CHANNEL_LAYERS = TEST_CHANNEL_LAYERS
-    set_up_data()
-    room = Room.objects.create()
-    user = get_user_model().objects.get(username="user0")
-    room.members.add(user)
-    room.save()
-    client = Client()
-    client.force_login(user=user)
+    client, room, user = prepare_room_and_user()
     communicator = WebsocketCommunicator(
-        application, f"/ws/chat/{room.id}/",
+        application, f"/ws/django_chatter/chatrooms/{room.id}/",
         headers=[
             (
                 b'cookie',
@@ -126,7 +119,7 @@ async def test_harmful_message_in_chat_consumer():
     connected, subprotocol = await communicator.connect()
     assert connected
     data = {
-        'type': 'text',
+        'message_type': 'text',
         'message': "<script>evil();</script>",
         'sender': user.username,
         'room_id': str(room.id),
@@ -137,3 +130,62 @@ async def test_harmful_message_in_chat_consumer():
     assert response['sender'] == "user0"
     assert response['room_id'] == str(room.id)
     await communicator.disconnect()
+
+@pytest.mark.asyncio
+@pytest.mark.django_db(transaction=True)
+async def test_chat_alert_consumer():
+    settings.CHANNEL_LAYERS = TEST_CHANNEL_LAYERS
+    client, room, user = prepare_room_and_user()
+    user1 = get_user_model().objects.get(username="user1")
+    other_client = Client()
+    other_client.force_login(user=user1)
+    room_with_two = Room.objects.create()
+    room_with_two.members.add(*[user, user1])
+    room_with_two.save()
+    room_with_user_1 = Room.objects.create()
+    room_with_user_1.members.add(user1)
+    room_with_user_1.save()
+
+    chat_communicator = WebsocketCommunicator(
+        application, f"/ws/django_chatter/chatrooms/{room_with_two.id}/",
+        headers=[
+            (
+                b'cookie',
+                f'sessionid={client.cookies["sessionid"].value}'.encode('ascii')
+            ),
+            (b'host', b'localhost:8000')]
+        )
+    connected, subprotocol = await chat_communicator.connect()
+    assert connected
+
+    user1_alert_communicator = WebsocketCommunicator(
+        application, f"/ws/django_chatter/users/{user1.username}/",
+        headers=[
+            (
+                b'cookie',
+                f'sessionid={other_client.cookies["sessionid"].value}'.encode('ascii')
+            ),
+            (b'host', b'localhost:8000')]
+    )
+    connected, subprotocol = await user1_alert_communicator.connect()
+    assert connected
+
+    data = {
+        'message_type': 'text',
+        'message': "<script>evil();</script>",
+        'sender': user.username,
+        'room_id': str(room_with_two.id),
+        }
+    await chat_communicator.send_json_to(data)
+    response = await chat_communicator.receive_json_from()
+    alert = await user1_alert_communicator.receive_json_from()
+    assert response['message'] == "&lt;script&gt;evil();&lt;/script&gt;"
+    assert response['sender'] == "user0"
+    assert response['room_id'] == str(room_with_two.id)
+    await chat_communicator.disconnect()
+
+
+    assert alert['message'] == "&lt;script&gt;evil();&lt;/script&gt;"
+    assert alert['sender'] == "user0"
+    assert alert['room_id'] == str(room_with_two.id)
+    await user1_alert_communicator.disconnect()